requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
idna==2.6
more-itertools==4.2.0
multidict==4.3.1
pluggy==0.6.0
py==1.5.3
pytest==3.6.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
vcrpy==1.12.0
wrapt==1.10.11
yarl==1.2.5
